+++ CHANGELOG.md
@@ -13,6 +13,8 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 * Add `soft_fill` to `Coords`, which aims to paint Gaussian blobs as coordinate
   locations.
 
+* Added `Boxes.bounding_box` method 
+
 * Added quantize method to bounding boxes which pushes the left-top coordinates
   to their floor and the right-bottom coordinates to their ceiling.
 
@@ -27,6 +29,8 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ### Changed
 
 * BREAKING: TLBR has been internally switched to LTRB 
+
+* Better Polygon coercion from geojson
 
 
 ## Version 0.6.10 - Released 2020-11-24